Create servo-media-streams library, move MediaStream stuff there
