+++ streams/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "servo-media-streams"
+version = "0.1.0"
+authors = ["Manish Goregaokar <manishsmail@gmail.com>"]
+edition = "2018"
+
+[dependencies]
