+++ webrtc/Cargo.toml
@@ -5,3 +5,6 @@ authors = ["Josh Matthews <josh@joshmatthews.net>"]
 
 [dependencies]
 boxfnonce = "0.1.0"
+
+[dependencies.servo-media-streams]
+path = "../streams"
