+++ webrtc/src/lib.rs
@@ -1,17 +1,13 @@
 extern crate boxfnonce;
+extern crate servo_media_streams;
+use servo_media_streams::MediaStream;
 
-use std::any::Any;
 use std::str::FromStr;
 
 use boxfnonce::SendBoxFnOnce;
 
 pub mod thread;
 pub mod capture;
-
-pub trait MediaStream: Any + Send {
-    fn as_any(&self) -> &Any;
-    fn as_mut_any(&mut self) -> &mut Any;
-}
 
 pub use thread::WebRtcController;
 
@@ -48,17 +44,6 @@ pub trait WebRtcSignaller: Send {
     fn on_negotiation_needed(&self, controller: &WebRtcController);
     fn close(&self);
     fn on_add_stream(&self, stream: Box<MediaStream>);
-}
-
-pub struct DummyMediaOutput;
-impl MediaOutput for DummyMediaOutput {
-    fn add_stream(&mut self, _stream: Box<MediaStream>) {}
-}
-
-/// This isn't part of the webrtc spec; it's a leaky abstaction while media streams
-/// are under development and example consumers need to be able to inspect them.
-pub trait MediaOutput: Send {
-    fn add_stream(&mut self, stream: Box<MediaStream>);
 }
 
 pub trait WebRtcBackend {
